requirements.txt
mysql-connector-python
python-dotenv
flask
flask_cors
PyJWT